requirements-dev.txt
black==24.4.2
mypy==1.11.0
ruff==0.5.6
pytest==8.2.2
pytest-cov==5.0.0
coverage==7.5.3
requests==2.32.3
grpcio==1.76.0
grpcio-tools==1.76.0
